Split general-purpose parsers from perseus-specific ones
